+++ docs/changelog.rst
@@ -5,6 +5,10 @@ CHANGELOG
 2.77.0+dev (XXXX-XX-XX)
 -----------------------
 
+**Minor improvements**
+
+- Show all infrastructures and signages on interventions
+
 **Bug fixes**
 
 - Show trail and path on intervention